fastapi_backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asn1crypto==1.5.1
cbor2==5.8.0
certifi==2026.2.25
certvalidator==0.11.1
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
fastapi==0.134.0
h11==0.16.0
helpers==0.2.0
idna==3.11
itsdangerous==2.2.0
oscrypto==1.3.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyOpenSSL==25.3.0
python-dotenv==1.2.2
requests==2.32.5
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
webauthn==2.7.1
python-jose[cryptography]==3.5.0
sqlalchemy==2.0.48